minor compilation warning fix
git-svn-id: svn://svn.berlios.de/open-iscsi@314 d7303112-9cec-0310-bdd2-e83a94d6c2b6

+++ usr/idbm.c
@@ -772,7 +772,7 @@ updated:
 	if (!passwd_done && !strcmp(#_param, name)) { \
 		passwd_done = 1; \
 		name = #_param "_length"; \
-		snprintf(passwd_len, 8, "%d", strlen(value)); \
+		snprintf(passwd_len, 8, "%d", (int)strlen(value)); \
 		value = passwd_len; \
 		goto setup_passwd_len; \
 	}
